[AIRFLOW-3459] Move DagPickle to separate file
