+++ airflow/__init__.py
@@ -26,6 +26,7 @@ in their PYTHONPATH. airflow_login should be based off the
 """
 from builtins import object
 from airflow import version
+from airflow import configuration
 from airflow.utils.log.logging_mixin import LoggingMixin
 
 __version__ = version.version
